requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=4.0.0
immanuel>=0.11.0
geopy>=2.3.0 